requirements.txt
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-embeddings-ollama
llama-index-vector-stores-postgres
psycopg2-binary~=2.9.10
dspy~=2.5.43
openpyxl~=3.1.5

numpy~=1.26.4
torch~=2.5.1
einops~=0.8.0
safetensors~=0.4.5
transformers~=4.47.1
llama-index-core~=0.12.9
sqlalchemy~=2.0.36
pandas~=2.2.3
pypdf~=5.1.0
ultralytics~=8.3.61
requests~=2.32.3
httpcore~=1.0.7
httpx~=0.27.2
pdfplumber~=0.11.5
fpdf~=1.7.2
matplotlib~=3.10.0
huggingface-hub~=0.27.0
pillow~=11.0.0
llmsherpa~=0.1.4